Implemented camera

+++ src/Window.c
@@ -32,7 +32,7 @@ void key_callback(GLFWwindow* window, int key, int scancode, int action, int mod
     	keys = 0;
 }
 
-void doMovement(float deltaTime) {
+void doMovement(double deltaTime) {
 	float deltaSpeed = 1.0;
 	if(keys == GLFW_KEY_W && actionPress == GLFW_PRESS)
         processKeyboard(FORWARD, deltaTime, deltaSpeed);
